fixing crashes and fixing some actions that were taking 0 time
