+++ src/main/java/pow/backend/action/BallSpell.java
@@ -88,6 +88,7 @@ public class BallSpell implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(actor));
         return ActionResult.failed(subactions);
     }
