+++ src/main/java/pow/backend/action/BoltSpell.java
@@ -60,6 +60,7 @@ public class BoltSpell implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(attacker));
         return ActionResult.failed(subactions);
     }
