+++ src/main/java/pow/backend/action/BreathSpell.java
@@ -65,6 +65,7 @@ public class BreathSpell implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(actor));
         return ActionResult.failed(subactions);
     }
