+++ src/main/java/pow/backend/action/CallPet.java
@@ -53,11 +53,12 @@ public class CallPet implements Action {
             subactions.add(new ShowEffect(new DungeonEffect(effectName, p)));
         }
 
-        subactions.add(new PhaseImpl(party.pet, targetLoc));
+        subactions.add(new PhaseImpl(actor, party.pet, targetLoc));
 
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(actor));
         return ActionResult.failed(subactions);
     }
