+++ src/main/java/pow/backend/action/ChainSpell.java
@@ -82,6 +82,7 @@ public class ChainSpell implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(attacker));
         return ActionResult.failed(subactions);
     }
