+++ src/main/java/pow/backend/action/CircleCut.java
@@ -65,6 +65,7 @@ public class CircleCut implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(attacker));
         return ActionResult.failed(subactions);
     }
