+++ src/main/java/pow/backend/action/GroupHeal.java
@@ -38,7 +38,7 @@ public class GroupHeal implements Action {
     }
 
     @Override
-    public boolean consumesEnergy() { return false; }
+    public boolean consumesEnergy() { return true; }
 
     private void healOne(Player player, int amount, GameBackend backend) {
         int desiredHealAmount = Math.max((int)Math.round(actor.getMaxHealth() * 0.01 * amount), amount);
