+++ src/main/java/pow/backend/action/Heal.java
@@ -35,5 +35,5 @@ public class Heal implements Action {
     }
 
     @Override
-    public boolean consumesEnergy() { return false; }
+    public boolean consumesEnergy() { return true; }
 }
