+++ src/main/java/pow/backend/action/QuakeSpell.java
@@ -53,6 +53,7 @@ public class QuakeSpell implements Action {
         // clear out last effect.
         // TODO: should this be new dungeonupdated?
         subactions.add(new ShowEffect(new DungeonEffect(Collections.emptyList())));
+        subactions.add(new CompletedAction(actor));
         return ActionResult.failed(subactions);
     }
