+++ src/main/java/pow/backend/action/Restore.java
@@ -35,5 +35,5 @@ public class Restore implements Action {
     }
 
     @Override
-    public boolean consumesEnergy() { return false; }
+    public boolean consumesEnergy() { return true; }
 }
